+++ libs/effects/src/index.ts
@@ -1,1 +1,7 @@
 export { rxEffect } from './lib/effects.functional';
+export type {
+  EffectCleanUpRef,
+  EffectsSetupFn,
+  RunOptions,
+  RxEffect,
+} from './lib/effects.functional';