+++ libs/effects/src/lib/effects-factory-function.spec.ts
@@ -103,7 +103,9 @@ class Service {
   triggerEffect(v: number) {
     return v;
   }
-  teardownEffect() {}
+  teardownEffect(): void {
+    return;
+  }
 }
 
 @Component({