+++ libs/effects/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## 1.3.0 (2024-07-10)
+
+* refactor(effects): rename runOnCleanUp to runOnInstanceDestroy ([8258f4f](https://github.com/michaelbe812/angular-kit-effects/commit/8258f4f))
+* refactor(effects): remove teardown callback on factory function, covered by runOnInstanceDestroy ([4b750f8](https://github.com/michaelbe812/angular-kit-effects/commit/4b750f8))
+* docs(effects): improve docs ([5cfcfbf](https://github.com/michaelbe812/angular-kit-effects/commit/5cfcfbf))
+
+## 1.2.0 (2024-03-28)
+
+* feat(rx-effects): introduce standalone effect functions ([c100a7f](https://github.com/michaelbe812/angular-kit-effects/commit/c100a7f))
+* chore(effects): setup semver ([97a3b14](https://github.com/michaelbe812/angular-kit-effects/commit/97a3b14))
+* docs(effects): update docs ([62db2a5](https://github.com/michaelbe812/angular-kit-effects/commit/62db2a5))
+
 ## 1.1.0 (2024-05-21)
 
 
@@ -9,6 +21,6 @@ This file was generated using [@jscutlery/semver](https://github.com/jscutlery/s
 
 ## 1.0.0 (2024-05-21)
 
-* feat(effects): add effects factory function to manage subs ([7254e4f](https://github.com/mikelgo/angular-kit/commit/7254e4f))
-* feat(effects): enhance options to pass a DestroyRef ([bcd1702](https://github.com/mikelgo/angular-kit/commit/bcd1702))
-* feat(effects): provide a clean up function for a single effect ([45c9ba3](https://github.com/mikelgo/angular-kit/commit/45c9ba3))
+* feat(effects): add effects factory function to manage subs ([7254e4f](https://github.com/michaelbe812/angular-kit-effects/commit/7254e4f))
+* feat(effects): enhance options to pass a DestroyRef ([bcd1702](https://github.com/michaelbe812/angular-kit-effects/commit/bcd1702))
+* feat(effects): provide a clean up function for a single effect ([45c9ba3](https://github.com/michaelbe812/angular-kit-effects/commit/45c9ba3))